Expense_Tracking_System/requirements.txt
django-crispy-forms==1.7.2
pkg-resources==0.0.0
